+++ src/lump.c
@@ -16,7 +16,7 @@ lump* read_directory(FILE* f,header* h){
     return directory;
 }
 
-int get_lump_index(lump* directory,char* lump_name,int lump_count){
+u16 get_lump_index(lump* directory,char* lump_name,int lump_count){
     for (int i = 0; i < lump_count;i++){
         if (strcmp(directory[i].lump_name,lump_name) == 0){
             return i;
